requirements.txt
Appium-Python-Client==4.0.1
attrs==23.2.0
behave==1.2.6
certifi==2024.7.4
charset-normalizer==3.3.2
h11==0.14.0
idna==3.7
loguru==0.7.2
outcome==1.3.0.post0
parse==1.20.2
parse-type==0.6.2
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
selenium==4.23.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
SpeechRecognition==3.10.4
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.9.0
urllib3==2.2.2
websocket-client==1.8.0
wsproto==1.2.0
